# source/packages/contracts/control/openapi.yaml
openapi: 3.1.0
info:
  title: Personal Agent Control API
  version: 0.3.0
  description: |
    Control-plane contract for daemon clients.
    Unknown `/v1/*` routes return `404` problem+json envelopes with correlation headers.
servers:
  - url: http://127.0.0.1:7071
    description: Localhost TCP default binding
security:
  - bearerAuth: []
paths:
...
  /v1/secrets/refs:
    post:
      operationId: upsertSecretReference
      summary: Upsert write-only secret reference metadata.
      parameters:
        - $ref: '#/components/parameters/CorrelationID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SecretReferenceUpsertRequest'
      responses:
        '200':
          description: Successful JSON response.
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationID'

            X-PersonalAgent-API-Version:
              $ref: '#/components/headers/APIVersion'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SecretReferenceResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '501':
          $ref: '#/components/responses/NotImplemented'
  /v1/secrets/refs/{workspace_id}/{name}:
    get:
      operationId: getSecretReference
      summary: Fetch secret reference metadata.
      parameters:
        - $ref: '#/components/parameters/WorkspaceID'
        - $ref: '#/components/parameters/SecretName'
        - $ref: '#/components/parameters/CorrelationID'
      responses:
        '200':
          description: Successful JSON response.
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationID'

            X-PersonalAgent-API-Version:
              $ref: '#/components/headers/APIVersion'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SecretReferenceResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '501':
          $ref: '#/components/responses/NotImplemented'
    delete:
      operationId: deleteSecretReference
      summary: Delete secret reference metadata.
      parameters:
        - $ref: '#/components/parameters/WorkspaceID'
        - $ref: '#/components/parameters/SecretName'
        - $ref: '#/components/parameters/CorrelationID'
      responses:
        '200':
          description: Successful JSON response.
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationID'

            X-PersonalAgent-API-Version:
              $ref: '#/components/headers/APIVersion'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SecretReferenceDeleteResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '501':
          $ref: '#/components/responses/NotImplemented'
...
  /v1/channels/twilio/webhook/serve:
    post:
      operationId: twilioWebhookServe
      summary: Start Twilio webhook server with optional cloudflared public tunnel.
      parameters:
        - $ref: '#/components/parameters/CorrelationID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TwilioWebhookServeRequest'
      responses:
        '200':
          description: Successful JSON response.
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationID'

            X-PersonalAgent-API-Version:
              $ref: '#/components/headers/APIVersion'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TwilioWebhookServeResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '501':
          $ref: '#/components/responses/NotImplemented'
...
  /v1/inspect/run:
    post:
      operationId: inspectRun
      summary: Inspect run details.
      parameters:
        - $ref: '#/components/parameters/CorrelationID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InspectRunRequest'
      responses:
        '200':
          description: Successful JSON response.
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationID'

            X-PersonalAgent-API-Version:
              $ref: '#/components/headers/APIVersion'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InspectRunResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '501':
          $ref: '#/components/responses/NotImplemented'
  /v1/inspect/transcript:
    post:
      operationId: inspectTranscript
      summary: Inspect transcript.
      parameters:
        - $ref: '#/components/parameters/CorrelationID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InspectTranscriptRequest'
      responses:
        '200':
          description: Successful JSON response.
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationID'

            X-PersonalAgent-API-Version:
              $ref: '#/components/headers/APIVersion'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InspectTranscriptResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '501':
          $ref: '#/components/responses/NotImplemented'
  /v1/inspect/memory:
    post:
      operationId: inspectMemory
      summary: Inspect memory entries.
      parameters:
        - $ref: '#/components/parameters/CorrelationID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InspectMemoryRequest'
      responses:
        '200':
          description: Successful JSON response.
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationID'

            X-PersonalAgent-API-Version:
              $ref: '#/components/headers/APIVersion'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InspectMemoryResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '501':
          $ref: '#/components/responses/NotImplemented'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: token
  parameters:
    CorrelationID:
      name: X-Correlation-ID
      in: header
      required: false
      schema:
        type: string
      description: Optional correlation ID propagated across daemon responses/events.
...
    WorkspaceID:
      name: workspace_id
      in: path
      required: true
      schema:
        type: string
    SecretName:
      name: name
      in: path
      required: true
      schema:
        type: string
  headers:
    CorrelationID:
      description: Correlation ID associated with this response.
      schema:
        type: string
    APIVersion:
      description: Daemon API contract version served by this endpoint.
      schema:
        type: string
        example: v1
...
  responses:
...
    BadRequest:
      description: Invalid request payload.
      headers:
        X-Correlation-ID:
          $ref: '#/components/headers/CorrelationID'
        X-PersonalAgent-API-Version:
          $ref: '#/components/headers/APIVersion'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Missing or invalid bearer token.
      headers:
        X-Correlation-ID:
          $ref: '#/components/headers/CorrelationID'
        X-PersonalAgent-API-Version:
          $ref: '#/components/headers/APIVersion'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Requested object not found.
      headers:
        X-Correlation-ID:
          $ref: '#/components/headers/CorrelationID'
        X-PersonalAgent-API-Version:
          $ref: '#/components/headers/APIVersion'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotImplemented:
      description: |
        Service route is registered but backing domain service is not configured.
        `error.code` is `service_not_configured` when applicable and `error.details`
        includes typed `service`, `domain`, and `remediation` metadata.
      headers:
        X-Correlation-ID:
          $ref: '#/components/headers/CorrelationID'
        X-PersonalAgent-API-Version:
          $ref: '#/components/headers/APIVersion'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    SecretReferenceUpsertRequest:
      type: object
      required:
        - workspace_id
        - name
        - backend
        - service
        - account
      properties:
        workspace_id:
          type: string
        name:
          type: string
        backend:
          type: string
        service:
          type: string
        account:
          type: string
    SecretReferenceRecord:
      type: object
      properties:
        workspace_id:
          type: string
        name:
          type: string
        backend:
          type: string
        service:
          type: string
        account:
          type: string
    SecretReferenceResponse:
      type: object
      properties:
        reference:
          $ref: '#/components/schemas/SecretReferenceRecord'
        correlation_id:
          type: string
    SecretReferenceDeleteResponse:
      type: object
      properties:
        reference:
          $ref: '#/components/schemas/SecretReferenceRecord'
        deleted:
          type: boolean
        correlation_id:
          type: string
...
    WorkflowRouteMetadata:
      type: object
      properties:
        available:
          type: boolean
        task_class:
          type: string
        provider:
          type: string
        model_key:
          type: string
        task_class_source:
          type: string
        route_source:
          type: string
        notes:
          type: string
...
    InspectRunRequest:
      type: object
      required:
        - run_id
      properties:
        run_id:
          type: string
    InspectTaskRun:
      type: object
      properties:
        run_id:
          type: string
        workspace_id:
          type: string
        task_id:
          type: string
        acting_as_actor_id:
          type: string
        state:
          type: string
        started_at:
          type: string
          format: date-time
        finished_at:
          type: string
          format: date-time
        last_error:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    InspectTask:
      type: object
      properties:
        task_id:
          type: string
        workspace_id:
          type: string
        requested_by_actor_id:
          type: string
        subject_principal_actor_id:
          type: string
        title:
          type: string
        description:
          type: string
        state:
          type: string
        priority:
          type: integer
        deadline_at:
          type: string
          format: date-time
        channel:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    InspectStep:
      type: object
      properties:
        step_id:
          type: string
        run_id:
          type: string
        step_index:
          type: integer
        name:
          type: string
        status:
          type: string
        interaction_level:
          type: string
        capability_key:
          type: string
        timeout_seconds:
          type: integer
        retry_max:
          type: integer
        retry_count:
          type: integer
        last_error:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    InspectRunArtifact:
      type: object
      properties:
        artifact_id:
          type: string
        run_id:
          type: string
        step_id:
          type: string
        artifact_type:
          type: string
        uri:
          type: string
        content_hash:
          type: string
        created_at:
          type: string
          format: date-time
    InspectAuditEntry:
      type: object
      properties:
        audit_id:
          type: string
        workspace_id:
          type: string
        run_id:
          type: string
        step_id:
          type: string
        event_type:
          type: string
        actor_id:
          type: string
        acting_as_actor_id:
          type: string
        correlation_id:
          type: string
        payload_json:
          type: string
        created_at:
          type: string
          format: date-time
    InspectRunResponse:
      type: object
      properties:
        task:
          $ref: '#/components/schemas/InspectTask'
        run:
          $ref: '#/components/schemas/InspectTaskRun'
        steps:
          type: array
          items:
            $ref: '#/components/schemas/InspectStep'
        artifacts:
          type: array
          items:
            $ref: '#/components/schemas/InspectRunArtifact'
        audit_entries:
          type: array
          items:
            $ref: '#/components/schemas/InspectAuditEntry'
        route:
          $ref: '#/components/schemas/WorkflowRouteMetadata'
    InspectTranscriptRequest:
      type: object
      properties:
        workspace_id:
          type: string
        thread_id:
          type: string
        limit:
          type: integer
    InspectTranscriptEvent:
      type: object
      properties:
        event_id:
          type: string
        workspace_id:
          type: string
        thread_id:
          type: string
        channel:
          type: string
        event_type:
          type: string
        direction:
          type: string
        assistant_emitted:
          type: boolean
        body_text:
          type: string
        sender_address:
          type: string
        occurred_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
    InspectTranscriptResponse:
      type: object
      properties:
        workspace_id:
          type: string
        events:
          type: array
          items:
            $ref: '#/components/schemas/InspectTranscriptEvent'
    InspectMemoryRequest:
      type: object
      properties:
        workspace_id:
          type: string
        owner_actor:
          type: string
        status:
          type: string
        limit:
          type: integer
    InspectMemoryItem:
      type: object
      properties:
        memory_id:
          type: string
        workspace_id:
          type: string
        owner_actor_id:
          type: string
        scope_type:
          type: string
        key:
          type: string
        status:
          type: string
        kind:
          type: string
        is_canonical:
          type: boolean
        token_estimate:
          type: integer
        source_summary:
          type: string
        source_count:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        value_json:
          type: string
    InspectMemoryResponse:
      type: object
      properties:
        workspace_id:
          type: string
        items:
          type: array
          items:
            $ref: '#/components/schemas/InspectMemoryItem'
...
    TwilioWebhookServeRequest:
      type: object
      properties:
        workspace_id:
          type: string
        listen_address:
          type: string
        signature_mode:
          type: string
        cloudflared_mode:
          type: string
        cloudflared_startup_timeout_ms:
          type: integer
          format: int64
        assistant_replies:
          type: boolean
        assistant_task_class:
          type: string
        assistant_system_prompt:
          type: string
        assistant_max_history:
          type: integer
        assistant_reply_timeout_ms:
          type: integer
          format: int64
        voice_response_mode:
          type: string
        voice_greeting:
          type: string
        voice_fallback:
          type: string
        sms_path:
          type: string
        voice_path:
          type: string
        run_for_ms:
          type: integer
          format: int64
    TwilioWebhookServeResponse:
      type: object
      properties:
        workspace_id:
          type: string
        signature_mode:
          type: string
        listen_address:
          type: string
        local_sms_webhook_url:
          type: string
        local_voice_webhook_url:
          type: string
        sms_webhook_url:
          type: string
        voice_webhook_url:
          type: string
        public_base_url:
          type: string
        cloudflared_mode:
          type: string
        cloudflared_available:
          type: boolean
        cloudflared_active:
          type: boolean
        cloudflared_dry_run:
          type: boolean
        cloudflared_binary_path:
          type: string
        assistant_replies:
          type: boolean
        assistant_task_class:
          type: string
        voice_response_mode:
          type: string
        warning:
          type: string
...
    ErrorResponse:
      type: object
      additionalProperties: true
      required:
        - error
        - correlation_id
        - type
        - title
        - status
        - detail
        - instance
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
            message:
              type: string
            details:
              description: |
                Optional domain-specific typed error metadata.
                For `service_not_configured` errors this includes `category`,
                `service` (`id`, `label`, `config_field`), `domain`, and `remediation`
                (`action`, `label`, `hint`).
        correlation_id:
          type: string
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string